Fix build failing on Linux because of an invalid cache path
TEMP variable may not be defined on Linux, and then the resulting path
was /platformio/build_cache, which is not directly on the root and
usually not writable. By adding /tmp/ prefix, this works on Linux where
the resulting path is /tmp/platformio/build_cache, and also on Windows,
where the path will be %TEMP%/tmp/platformio/build_cache, which is also
fine.